% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{Geweke_convergence}
\alias{Geweke_convergence}
\title{Geweke Convergence}
\usage{
Geweke_convergence(THETA)
}
\arguments{
\item{THETA}{Matrix of parameter estimates from IRTM}
}
\value{
Proportion of values that fail the Geweke convergence test (p < 0.05) for each parameter
}
\description{
Runs Geweke tests to assess MCMC convergence
}
