% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_IRT.R
\name{M_constrained_irt_continuous}
\alias{M_constrained_irt_continuous}
\title{Constrained IRT model for continuous outcomes}
\usage{
M_constrained_irt_continuous(
  Y,
  d,
  M = NULL,
  theta_fix = NULL,
  which_fix = NULL,
  nburn = 1000,
  nsamp = 1000,
  thin = 10,
  learn_Sigma = TRUE,
  learn_Omega = FALSE,
  hyperparameters = list(),
  display_progress = TRUE
)
}
\arguments{
\item{Y}{a N x K matrix of responses given by N respondents to K items. Can contain missing values.}

\item{d}{an integer specifying the number of latent dimensions.}

\item{M}{a list of K d x d matrices (default=NULL).}

\item{theta_fix}{a matrix with d columns containing the values of the latent dimensions for respondents that have pre-specified latent factors.}

\item{which_fix}{a vector containing the indices of the respondents for which latent factors have been fixed.}

\item{nburn}{an integer specifying the number of burn-in MCMC iterations.}

\item{nsamp}{an integer specifying the number of sampling MCMC iterations.}

\item{thin}{an integer specifying the number of thinning MCMC samples.}

\item{learn_Sigma}{a Boolean specifying whether a covariance matrix for the latent factors should be learned.}

\item{learn_Omega}{a Boolean specifying whether a covariance matrix for the latent loadings should be learned.}

\item{hyperparameters}{a list of hyperparameters for the model.}

\item{display_progress}{a Boolean specifying whether a progress bar should be displayed.}
}
\value{
A list with posterior samples for \code{lambda}, \code{b},
\code{theta}, and optionally \code{Sigma} and \code{Omega}, in the
same format as \code{M_constrained_irt()}.
}
\description{
This fits the same constrained multidimensional IRT model as
\code{M_constrained_irt()}, but assumes the observed responses
\code{Y} are continuous and directly follow a Gaussian measurement
model with mean \eqn{\lambda_k^\top \theta_i - b_k}.
}
