% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_IRT.R
\name{M_constrained_irt_family}
\alias{M_constrained_irt_family}
\title{Wrapper for IRT-M binary or continuous}
\usage{
M_constrained_irt_family(
  Y,
  d,
  family = c("binary", "continuous"),
  M = NULL,
  theta_fix = NULL,
  which_fix = NULL,
  nburn = 1000,
  nsamp = 1000,
  thin = 10,
  learn_Sigma = TRUE,
  learn_Omega = FALSE,
  hyperparameters = list(),
  display_progress = TRUE
)
}
\arguments{
\item{Y}{a N x K matrix of responses given by N respondents to K items. Can contain missing values.}

\item{d}{an integer specifying the number of latent dimensions.}

\item{family}{Character string specifying the outcome type:
\code{"binary"} or \code{"continuous"}.}

\item{M}{a list of K d x d matrices (default=NULL).}

\item{theta_fix}{a matrix with d columns containing the values of the latent dimensions for respondents that have pre-specified latent factors.}

\item{which_fix}{a vector containing the indices of the respondents for which latent factors have been fixed.}

\item{nburn}{an integer specifying the number of burn-in MCMC iterations.}

\item{nsamp}{an integer specifying the number of sampling MCMC iterations.}

\item{thin}{an integer specifying the number of thinning MCMC samples.}

\item{learn_Sigma}{a Boolean specifying whether a covariance matrix for the latent factors should be learned.}

\item{learn_Omega}{a Boolean specifying whether a covariance matrix for the latent loadings should be learned.}

\item{hyperparameters}{a list of hyperparameters for the model.}

\item{display_progress}{a Boolean specifying whether a progress bar should be displayed.}
}
\value{
A list with posterior samples for \code{lambda}, \code{b},
\code{theta}, and optionally \code{Sigma} and \code{Omega}, in the
same format as \code{M_constrained_irt()}.
}
\description{
Convenience wrapper that dispatches to either
}
\details{
\code{M_constrained_irt()} (binary probit) or
\code{M_constrained_irt_continuous()} (Gaussian) depending on
\code{family}.
}
\examples{
\dontrun{
  # Binary outcomes
  fit_bin <- M_constrained_irt_family(Y_bin, d = 2, family = "binary")

  # Continuous outcomes
  fit_cont <- M_constrained_irt_family(Y_cont, d = 2, family = "continuous")
}

}
