% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mse}
\alias{mse}
\title{Mean Squared Error}
\usage{
mse(ytrue, ypred, aggregate = TRUE, root = FALSE)
}
\arguments{
\item{ytrue}{observed values,}

\item{ypred}{predicted values}

\item{aggregate}{logical for whether to take mean of estimate}

\item{root}{logical for whether to return square root of MSE}
}
\value{
mean squared error
}
\description{
loss function for benchmarks
}
