% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{S109_rollcalls}
\alias{S109_rollcalls}
\title{Roll Call Votes for S109 Legislative Session}
\format{
A data frame with the following variables:
\describe{
\item{congress}{Congressional session number (109)}
\item{chamber}{Chamber ("Senate")}
\item{rollnumber}{Identifying number for the vote}
\item{date}{Date when the vote occurred}
\item{session}{Session (1 or 2)}
\item{clerk_rollnumber}{Roll call number assigned by the Clerk of the House or Secretary of the Senate}
\item{yea_count}{Count of "Yea" affirmative votes}
\item{nay_count}{Count of "Nay" negative votes}
\item{nominate_mid_1}{Midpoint estimate for the first NOMINATE dimension}
\item{nominate_mid_2}{The midpoint estimate for the second NOMINATE dimension}
\item{nominate_spread_1}{Spread parameter for the first NOMINATE dimension}
\item{nominate_spread_2}{Spread parameter for the second NOMINATE dimension}
\item{nominate_log_likelihood}{Log-likelihood of the NOMINATE estimates}
\item{bill_number}{An identifying number of the bill or resolution being voted on}
\item{vote_result}{Vote outcome (e.g., "Passed" or "Failed")}
\item{vote_desc}{Text description of the substance of the vote}
\item{vote_question}{Parliamentary question being voted on (e.g., "On Passage", "On Motion to Suspend the Rules")}
\item{dtl_desc}{Detailed description of the bill, when avaliable}
}
}
\source{
PSCL package:
Jackman S (2024). pscl: Classes and Methods for R Developed in the Political Science Computational Laboratory.
University of Sydney, Sydney, Australia. R package version 1.5.9, https://github.com/atahk/pscl/.
}
\description{
Record of roll call votes during the S109 legislative session.
Each vote includes the NOMINATE scores associated with the vote
}
\details{
Datasets for IRT-M Package

Datasets for constrained IRT examples.

The S109 data is used in:
Morucci, Marco, Margaret J. Foster, Kaitlyn Webster, So Jin Lee, and David A. Siegel.
"Measurement That Matches Theory: Theory-Driven Identification in Item Response Theory Models."
American Political Science Review (2024): 1-19.
}
\keyword{datasets}
