% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{standardize_theta}
\alias{standardize_theta}
\title{Standardize Theta}
\usage{
standardize_theta(theta, Sigma)
}
\arguments{
\item{theta}{estimated object}

\item{Sigma}{covariance matrix}
}
\value{
theta divided by sigma param
}
\description{
standardizes theta estimates
}
