% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synth_idvs}
\alias{synth_idvs}
\title{Synthetic Independent Variables}
\format{
A data frame with synthetic research variables
}
\description{
Synthetic dataset of independent variables for statistical analysis.
}
\keyword{datasets}
