% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synth_questions}
\alias{synth_questions}
\title{Questions for the Synthetic European sentiment survey in the vignette}
\format{
A data frame with 3000 rows and 148 columns representing synthetic survey responses
}
\description{
A synthetic dataset with 3000 rows and 148 questions.
This data replicates the structure of questions following Eurobarometer 94.3 (2021).
It is not intended to be analyzed independently from the vignette.
'
}
\keyword{datasets}
