% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_analysis.R
\name{theta_av}
\alias{theta_av}
\title{Average Thetas}
\usage{
theta_av(theta_array)
}
\arguments{
\item{theta_array}{An array of dimension (N x d x nsamp/thin) containing posterior samples of respondent latent trait values}
}
\value{
theta_av N x d matrix of average thetas
}
\description{
Compute matrix of theta means over posterior distributions
}
