% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{theta_lambda_traceplots}
\alias{theta_lambda_traceplots}
\title{Theta Lambda Traceplots}
\usage{
theta_lambda_traceplots(irt, i = NULL, k = NULL)
}
\arguments{
\item{irt}{An object containing theta and lambda parameters from an IRTM model}

\item{i}{Index of the respondent to plot (randomly selected if NULL)}

\item{k}{Index of the item to plot (randomly selected if NULL)}
}
\value{
Plots of theta, lambda, and their product across MCMC iterations
}
\description{
Creates traceplots for IRT parameter convergence diagnostics
}
