% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MCodes}
\alias{MCodes}
\title{Methodological Codes}
\format{
A data frame or other appropriate data structure
}
\description{
Coding scheme and reference codes used in the legislative analysis.
}
\keyword{datasets}
