% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{S109_votes}
\alias{S109_votes}
\title{Comprehensive Voting Data}
\format{
Similar documentation as above...
}
\description{
Aggregated voting information for the S109 legislative session.
}
\keyword{datasets}
