% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synth_questions}
\alias{synth_questions}
\title{Synthetic Survey Questions}
\format{
A data frame containing survey question data
}
\description{
Synthetic dataset of survey questions for research purposes.
}
\keyword{datasets}
